+++ src/utils/buildSchema.ts
@@ -1,4 +1,5 @@
 import { GraphQLSchema } from "graphql";
+import { Options as PrintSchemaOptions } from "graphql/utilities/schemaPrinter";
 
 import { SchemaGenerator, SchemaGeneratorOptions } from "../schema/schema-generator";
 import { loadResolversFromGlob } from "../helpers/loadResolversFromGlob";
@@ -6,20 +7,29 @@ import {
   defaultSchemaFilePath,
   emitSchemaDefinitionFileSync,
   emitSchemaDefinitionFile,
+  defaultPrintSchemaOptions,
 } from "./emitSchemaDefinitionFile";
+
+interface EmitSchemaFileOptions extends PrintSchemaOptions {
+  path?: string;
+}
 
 export interface BuildSchemaOptions extends SchemaGeneratorOptions {
   /** Array of resolvers classes or glob paths to resolver files */
   resolvers: Array<Function | string>;
-  /** Path to the file to where emit the schema or `true` for the default `./schema.gql` one */
-  emitSchemaFile?: string | boolean;
+  /**
+   * Path to the file to where emit the schema
+   * or config object with print schema options
+   * or `true` for the default `./schema.gql` one
+   */
+  emitSchemaFile?: string | boolean | EmitSchemaFileOptions;
 }
 export async function buildSchema(options: BuildSchemaOptions): Promise<GraphQLSchema> {
   loadResolvers(options);
   const schema = await SchemaGenerator.generateFromMetadata(options);
   if (options.emitSchemaFile) {
-    const schemaFileName = getSchemaDefinitionFileName(options);
-    await emitSchemaDefinitionFile(schemaFileName, schema);
+    const { schemaFileName, printSchemaOptions } = getEmitSchemaDefinitionFileOptions(options);
+    await emitSchemaDefinitionFile(schemaFileName, schema, printSchemaOptions);
   }
   return schema;
 }
@@ -28,8 +38,8 @@ export function buildSchemaSync(options: BuildSchemaOptions): GraphQLSchema {
   loadResolvers(options);
   const schema = SchemaGenerator.generateFromMetadataSync(options);
   if (options.emitSchemaFile) {
-    const schemaFileName = getSchemaDefinitionFileName(options);
-    emitSchemaDefinitionFileSync(schemaFileName, schema);
+    const { schemaFileName, printSchemaOptions } = getEmitSchemaDefinitionFileOptions(options);
+    emitSchemaDefinitionFileSync(schemaFileName, schema, printSchemaOptions);
   }
   return schema;
 }
@@ -45,8 +55,22 @@ function loadResolvers(options: BuildSchemaOptions) {
   });
 }
 
-function getSchemaDefinitionFileName(options: BuildSchemaOptions): string {
-  return typeof options.emitSchemaFile === "string"
-    ? options.emitSchemaFile
-    : defaultSchemaFilePath;
+function getEmitSchemaDefinitionFileOptions(
+  buildSchemaOptions: BuildSchemaOptions,
+): {
+  schemaFileName: string;
+  printSchemaOptions: PrintSchemaOptions;
+} {
+  return {
+    schemaFileName:
+      typeof buildSchemaOptions.emitSchemaFile === "string"
+        ? buildSchemaOptions.emitSchemaFile
+        : typeof buildSchemaOptions.emitSchemaFile === "object"
+          ? buildSchemaOptions.emitSchemaFile.path || defaultSchemaFilePath
+          : defaultSchemaFilePath,
+    printSchemaOptions:
+      typeof buildSchemaOptions.emitSchemaFile === "object"
+        ? buildSchemaOptions.emitSchemaFile
+        : defaultPrintSchemaOptions,
+  };
 }